\name{practprepr-package}
\alias{practprepr-package}
\alias{practprepr}
\docType{package}
\title{practprepr: Paquete de practica}
\description{
Paquete para aprender como se crean, se distribuyen y se usan en R
Contiene dos funciones sencillas
}
\details{
\tabular{ll}{
Package: \tab practprepr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-12-21\cr
}
}
\author{javierph}
\keyword{ package }
\section{practprepr functions}{

suma.dos.numeros
resta.dos.numeros
}
}
