% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/practprepr.R
\docType{package}
\name{practprepr}
\alias{practprepr}
\alias{practprepr-package}
\title{practprepr: Paquete de practica}
\description{
Paquete para aprender como se crean, se distribuyen y se usan en R
Contiene dos funciones sencillas
}
\section{practprepr functions}{

suma.dos.numeros
resta.dos.numeros
}

