% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Suma los numeros pasados por parametro}
\usage{
suma.dos.numeros(x = 0, y = 0)
}
\arguments{
\item{x}{Primer numero a sumar. Es 0 por defecto}

\item{y}{Segundo numero a sumar. Es 0 por defecto}
}
\value{
La suma de \code{x} e \code{y}
}
\description{
Suma los numeros pasados por parametro
}
\examples{
suma.dos.numeros(13, 31)
suma.dos.numeros(-1, 3.14159)
}

