% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{resta.dos.numeros}
\alias{resta.dos.numeros}
\title{Resta los numeros pasados por parametro}
\usage{
resta.dos.numeros(x, y)
}
\arguments{
\item{x}{Primer numero a restar}

\item{y}{Segundo numero a restar}
}
\value{
La resta de \code{x} menos \code{y}
}
\description{
Resta los numeros pasados por parametro
}
\examples{
resta.dos.numeros(13, 31)
resta.dos.numeros(-1, 3.14159)
}

